Работа над замечаниями: - разложил классы по пакетам, выделив pojo - вынес в переменную части Урлов - пересоздал алюр
