+++ src/main/java/yandex/client/RestClient.java
@@ -1,4 +1,4 @@
-package yandex.example;
+package yandex.client;
 
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
@@ -11,7 +11,8 @@ import static io.restassured.http.ContentType.JSON;
 public class RestClient {
 
     public static final String BASE_URL = "https://stellarburgers.nomoreparties.site/api/";
-    protected RequestSpecification getBaseSpec(){
+
+    protected RequestSpecification getBaseSpec() {
         return new RequestSpecBuilder()
                 .setContentType(JSON)
                 .setBaseUri(BASE_URL)
